Free_Data_API-feat-microservice-refactor/panoramic-market-analyzer/requirements.txt
pyyaml
pandas
numpy
yfinance
requests
requests-cache
fredapi
duckdb
